UI/requirements.txt
Flask==2.3.3
cx_Oracle==8.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7